+++ doorstop/__init__.py
@@ -10,6 +10,6 @@ VERSION = __project__ + '-' + __version__
 try:
     from doorstop.common import DoorstopError, DoorstopWarning, DoorstopInfo
     from doorstop.core import Item, Document, Tree
-    from doorstop.core import build, report, find_document, find_item
+    from doorstop.core import importer, build, report, find_document, find_item
 except ImportError:  # pragma: no cover, manual test
     pass